requirements.txt
amqp
APScheduler
asgiref
backports.zoneinfo==0.2.1;python_version<"3.9"
cachetools
celery
django-celery-beat==2.4.0
certifi
cffi
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
cryptography
dj-database-url
Django
django-environ
django-filter
django-heroku
django-model-utils
django-mysql
gunicorn
HireFire
idna
kombu
mysqlclient
ndg-httpsclient
oauthlib
prompt-toolkit
psycopg2-binary
ptb
pyasn1
pycparser
Pygments
pyOpenSSL
PySocks
python-telegram-bot==13.7
pytz
PyYAML
Redis
requests
requests-oauthlib
searchtweets
six
sqlparse
tornado
tweepy==4.12.1
tweet-parser
tzlocal
urllib3
vine
wcwidth
whitenoise
